Implement timeline view on homepage
This commit enhances the homepage by introducing a timeline which displays Food, Symptom and Movement logs for the given day in chronological order.

This feature provides the users with a convenient overview of the day's events, and allows for the identification of correlations between the different types of data.

Changes made:
- Introduced a `Log` interface to be able to unify different log types within a single collection
- Added repository methods to retrieve logs within a specified time range
- Implemented timeline functionality on the HomeScreen