Add view food log page
A view food log page has been introduced which provides users with detailed information for a specific food log entry. Users can access this page by clicking on a food log item, either via the home or logs page.

This view page serves as an entry point for future enhancements, including the addition of edit and delete functionality for the logs.

Other changes:
- Adds onClick support for a LogItemCard, enabling navigation to the view page upon clicking the card