Add data models for food-related entities
This commit introduces separate data models for food-related entities to decouple the database schema from the rest of the application logic.

This change fosters better abstraction and separation of concern between the storage solution and business logic and UI layers.  This allows for easier modification of the database schema or storage solution, as well as promoting clean architecture.