Change Log from an interface to a sealed class
Now that all log models have their `date` field at the top-level, there is no longer the need to have an interface to define how to get the date for a given log type.

Instead we can more easily use a data class to represent a shared Log class with a `date` field.