Close dropdown when creating item
With the current behaviour, when a user chooses to create a new item from the search dropdown the dropdown remains expanded. This requires an unnecessary additional click to then close the dropdown to add the item to a given log.

This commit fixes this by closing the dropdown as part of the onClick handler for creating a new item.