Improve search results by prioritising items starting with the search term
When adding items to a food log it can be frustrating if the desired item is not near the top of the list, especially if it stats with the search term. This issue occurs because the current search approach simply checks if the items contain the search term.

To address this, this commit introduces a new prioritisation strategy that groups items starting with the search term before those that merely contain it. This enhances the usability and speeds up the selection process.

Opted for a straightforward prefix prioritisation instead of a more complex algorithm (e.g. Levenshtein distance) as the typical datasets are reliably small and simple.