Add ability to copy a food log
When similar meals are eaten on repeat it can be time consuming to re-log the items instead of being able to copy them.

This commit addresses the above by adding copy functionality when viewing a food log. Clicking the copy button navigates the user to the add food log page with the selected items prefilled from the log being copied, the date and time are not copied as it is likely a recent meal being logged.

Additional changes:
- The logic to determine when to navigate from the view food log screen as been moved into the view model for better separation of business logic