Fix FAB insets and update behaviour on Logs Screen
Due to nested Scaffolds, the FAB on pages with a NavigationBar were positioned much higher than they should have been. This is fixed by removing the WindowInsets on those screens.

Also, the FAB behaviour within the Logs Screen was no longer consistent with the Home Screen when logging a food item with the Mealie integration enabled, as it opened a bottom sheet instead of a FAB menu. This is also updated.