Adds the ability to log drinks
Now that the home page is able to display any drink logs that exist in the database. This commit adds the ability to log drinks via the homepage FAB.

The log page works in the same way as adding a food log in that a user can search for existing drinks to log or create new ones.