Use home page date when adding logs
When navigating to a different date on the home page and choosing to add a log, the date on the log defaults to the current date. This doesn't follow the expected behaviour of the log being added to the date being viewed, and means there are additional steps required to correct it.

This commit adds the date as a navigation prop to all add log routes to allow the home page date to initialise the appropriate view models.

It also extracts common navigation arguments into a separate class for re-use across features.